builder/Dockerfile: suppress InvalidDefaultArgInFrom warning emitted by `docker build` when running `make release`
Recent versions of `docker build` started generating the InvalidDefaultArgInFrom warning if Dockerfile contains
an ARG without default value. While this warning doesn't affect running `make release`,
it is better suppressing the warning, so it doesn't clutter `make release` output with the noise,
which can hide real issues in the future.